chore: rename modal folder
