+++ src/bridges/ethereum-darwinia/Darwinia2Ethereum.tsx
@@ -10,8 +10,8 @@ import Web3 from 'web3';
 import { AssetGroup } from '../../components/form-control/AssetGroup';
 import { PolkadotAccountsItem } from '../../components/form-control/PolkadotAccountsItem';
 import { RecipientItem } from '../../components/form-control/RecipientItem';
-import { TransferConfirm } from '../../components/modal/TransferConfirm';
-import { TransferSuccess } from '../../components/modal/TransferSuccess';
+import { TransferConfirm } from '../../components/tx/TransferConfirm';
+import { TransferSuccess } from '../../components/tx/TransferSuccess';
 import { FORM_CONTROL } from '../../config/constant';
 import { getToken, useAfterTx, useApi, useDarwiniaAvailableBalances, useDeparture, useTx } from '../../hooks';
 import { AvailableBalance, CrossChainComponentProps, CrossChainPayload, DarwiniaAsset, Token } from '../../model';
