+++ src/bridges/substrate-substrateDVM/Substrate2SubstrateDVM.tsx
@@ -42,8 +42,8 @@ import { Balance } from '../../components/form-control/Balance';
 import { MaxBalance } from '../../components/form-control/MaxBalance';
 import { PolkadotAccountsItem } from '../../components/form-control/PolkadotAccountsItem';
 import { RecipientItem } from '../../components/form-control/RecipientItem';
-import { TransferConfirm } from '../../components/modal/TransferConfirm';
-import { TransferSuccess } from '../../components/modal/TransferSuccess';
+import { TransferConfirm } from '../../components/tx/TransferConfirm';
+import { TransferSuccess } from '../../components/tx/TransferSuccess';
 import { IssuingSubstrateTxPayload, Substrate2SubstrateDVMPayload } from './model';
 import { issuing } from './utils/tx';
