+++ src/components/tx/ApproveConfirm.tsx
@@ -1,16 +1,16 @@
 import { useTranslation } from 'react-i18next';
 import { CrossChainPayload } from '../../model';
-import { Des } from './Des';
+import { IDescription } from '../widget/IDescription';
 
 export function ApproveConfirm({ value }: { value: CrossChainPayload<{ sender: string }> }) {
   const { t } = useTranslation();
 
   return (
-    <Des
+    <IDescription
       title={
         <span className="capitalize">{t('{{network}} Network Address', { network: value.direction.from?.name })}</span>
       }
       content={value.sender}
-    ></Des>
+    ></IDescription>
   );
 }
